Pagelogging service -> ticketService
