+++ server/services/ticket.js
@@ -1,5 +1,6 @@
 import Joi from 'joi';
-import Ticket from '../models/ticket';
+import _ from 'lodash';
+import Ticket, { actionTypes } from '../models/ticket';
 import JoiHelper from '../helpers/JoiHelper';
 import alertService from './alert';
 
@@ -41,10 +42,33 @@ function getTicketsByDate(filterOpts) {
   return Ticket.getTicketsByDate(filterOpts);
 }
 
+function addAction(ticketId, actionType, userId) {
+  const schema = {
+    ticketId: Joi.string().hex().length(24).required(),
+    actionType: Joi.any().only(_.keys(actionTypes)).required(),
+    userId: Joi.string().hex().length(24)
+  };
+  return JoiHelper.validate({ ticketId, actionType, userId }, schema)
+    .then(values => Ticket.addAction(values.ticketId, values.actionType, values.userId));
+}
+
+function removeAction(ticketId, actionType, timestamp, userId) {
+  const schema = {
+    ticketId: Joi.string().hex().length(24).required(),
+    actionType: Joi.any().only(_.keys(actionTypes)).required(),
+    timestamp: Joi.number(),
+    userId: Joi.string().hex().length(24)
+  };
+  return JoiHelper.validate({ ticketId, actionType, timestamp, userId }, schema)
+    .then(v => Ticket.removeAction(v.ticketId, v.actionType, v.timestamp, v.userId));
+}
+
 export default {
   createTicket,
   getById,
   updateTicket,
   deleteById,
-  getTicketsByDate
+  getTicketsByDate,
+  addAction,
+  removeAction
 };
