+++ server/tests/services/ticket.test.js
@@ -1,9 +1,11 @@
 import httpStatus from 'http-status';
 import ticketService from '../../services/ticket';
+import Ticket, { actionTypes } from '../../models/ticket';
 import { build, fixtures } from '../../utils/factories';
 
 describe('## Ticket Service', () => {
   const ticketObject = fixtures.ticket();
+  const userId = '123456789098765432123456';
 
   describe('# createTicket()', () => {
     context('with valid ticket details', () => {
@@ -229,6 +231,95 @@ describe('## Ticket Service', () => {
         ticketService.getTicketsByDate(filters)
           .then((tickets) => {
             expect(tickets).to.have.lengthOf(4);
+            done();
+          });
+      });
+    });
+  });
+
+  describe('# addAction()', () => {
+    let ticketId;
+
+    beforeEach((done) => {
+      build('ticket', fixtures.ticket())
+        .then((ticket) => {
+          ticketId = ticket.id;
+          done();
+        });
+    });
+
+    context('with valid input parameters', () => {
+      it('adds an action successfully', (done) => {
+        ticketService.addAction(ticketId, actionTypes.CREATED, userId)
+          .then((ticket) => {
+            expect(ticket.actions[0].actionTaken).to.equal(actionTypes.CREATED);
+            expect(ticket.actions[0].userId.toString()).to.equal(userId);
+            done();
+          });
+      });
+    });
+
+    context('with invalid input parameters', () => {
+      it('fails to add an action', (done) => {
+        ticketService.addAction(ticketId, 'testing', userId)
+          .catch((err) => {
+            expect(err.details[0].message)
+              .to.equal('"actionType" must be one of [CREATED, PAGE_SENT, ACKNOWLEDGED, REJECTED, CLOSED]');
+            done();
+          });
+      });
+    });
+  });
+
+  describe('# removeAction()', () => {
+    let ticketId;
+
+    const action = {
+      actionTaken: actionTypes.CREATED,
+      userId,
+      timestamp: Date.now()
+    };
+
+    beforeEach((done) => {
+      const updates = { $set: { actions: [action] } };
+
+      build('ticket', fixtures.ticket())
+        .then((ticket) => {
+          ticketId = ticket.id;
+          Ticket.findByIdAndUpdate(ticketId, updates, { new: true }, (err, ticket) => {
+            if (ticket) {
+              done();
+            }
+          });
+        });
+    });
+
+    context('with valid input parameters', () => {
+      it('removes an action successfully', (done) => {
+        ticketService.removeAction(ticketId, action.actionTaken, action.timestamp, action.userId)
+          .then((ticket) => {
+            expect(ticket.actions).to.be.empty;
+            done();
+          });
+      });
+
+      it('doesn\'t remove actions that don\'t match the search criteria', (done) => {
+        ticketService.removeAction(ticketId, actionTypes.CLOSED, action.timestamp, action.userId)
+          .then((ticket) => {
+            expect(ticket.actions).to.have.length(1);
+            expect(ticket.actions[0].actionTaken).to.equal(actionTypes.CREATED);
+            expect(ticket.actions[0].userId.toString()).to.equal(userId);
+            done();
+          });
+      });
+    });
+
+    context('with invalid input parameters', () => {
+      it('fails to remove an action', (done) => {
+        ticketService.removeAction(ticketId, 'testing', action.timestamp)
+          .catch((err) => {
+            expect(err.details[0].message)
+              .to.equal('"actionType" must be one of [CREATED, PAGE_SENT, ACKNOWLEDGED, REJECTED, CLOSED]');
             done();
           });
       });
